requirements.txt
web3==6.11.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pytorch==2.1.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
cryptography==41.0.4
aiohttp==3.8.6
pytest==7.4.3
black==23.10.1
flake8==6.1.0